requirements.txt
flask
joblib
matplotlib
numpy
pandas
requests
websocket-client
scikit-learn
scipy
